fixed celery's decorator import bug
